pyproject.toml
[project]
name = "proxy-voter"
version = "0.1.0"
description = "Automated proxy voting system"
requires-python = ">=3.13"
dependencies = [
    "anthropic",
    "aiosqlite",
    "beautifulsoup4",
    "fastapi",
    "lxml",
    "playwright",
    "pydantic-settings",
    "python-multipart",
    "resend",
    "uvicorn[standard]",
]

[project.optional-dependencies]
dev = [
    "httpx",
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
